Added implementation file for Position class
