+++ CMakeLists.txt
@@ -17,5 +17,6 @@ target_sources(
   Sources/Globals.cpp
   Sources/Texture.cpp
   Sources/Timer.cpp
+  Sources/Position.cpp
 )
 target_link_libraries(pacman_sdl ${SDL2_LIBRARIES} SDL2::Main SDL2::Image SDL2::TTF SDL2::Mixer)
