+++ Headers/Position.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 class Position{
 	public:
 		Position(short x = 0, short y = 0);
@@ -14,39 +16,3 @@ class Position{
 	private:
 		short x, y;
 };
-
-Position::Position(short x, short y){
-	this->x = x;
-	this->y = y;
-}
-
-short Position::GetX(){
-	return this->x;
-}
-
-short Position::GetY(){
-	return this->y;
-}
-
-Position Position::GetPos(){
-	Position MyPos(this->x, this->y);
-	return MyPos;
-}
-
-void Position::ModX(short NewX){
-	this->x = NewX;
-}
-
-void Position::ModY(short NewY){
-	this->y = NewY;	
-}
-
-void Position::ModCoords(short NewX, short NewY){
-	this->x = NewX;
-	this->y = NewY;
-}
-
-void Position::ModPos(Position NewPos){
-	this->x = NewPos.GetX();
-	this->y = NewPos.GetY();
-}
